lesson2/requirements.txt
chardet==4.0.0
pyyaml==6.0